Streamline various build things. [publish windows-cpp11]

+++ scripts/build-atom.sh
@@ -72,7 +72,11 @@ mv $build_dir/version $build_dir/version.txt
 mv $build_dir/LICENSE $build_dir/LICENSE.txt
 
 echo "running npm install"
-BUILD_PLATFORM=$platform npm install --production
+BUILD_PLATFORM=$platform npm install --production \
+--target_platform=$platform \
+--target=$node_version \
+--target_arch=$arch \
+--fallback-to-build=false
 
 # Remove extra deps dirs to save space
 deps="node_modules/mbtiles/node_modules/sqlite3/deps
@@ -82,22 +86,6 @@ for depdir in $deps; do
     rm -r $app_dir/$depdir
 done
 
-# Go through pre-gyp modules and rebuild for target platform/arch.
-modules="node_modules/mapnik
-node_modules/mbtiles/node_modules/sqlite3
-node_modules/mapnik-omnivore/node_modules/srs
-node_modules/mapnik-omnivore/node_modules/gdal"
-for module in $modules; do
-    rm -r $app_dir/$module/lib/binding
-    cd $app_dir/$module
-    echo "resbuilding $app_dir for $platform $node_version $arch"
-    $app_dir/node_modules/.bin/node-pre-gyp install \
-        --target_platform=$platform \
-        --target=$node_version \
-        --target_arch=$arch \
-        --fallback-to-build=false
-done
-
 cd /tmp
 
 # win32: installer using nsis
@@ -105,35 +93,17 @@ if [ $platform == "win32" ]; then
     if ! which makensis > /dev/null; then echo "makensis command not found"; exit 1; fi;
     if ! which signcode > /dev/null; then echo "signcode command not found"; exit 1; fi;
     if ! which expect > /dev/null; then echo "expect command not found"; exit 1; fi;
+    if ! which windowsign > /dev/null; then echo "windowsign command not found"; exit 1; fi;
 
     # windows code signing
-    # uses mono `signcode` and `expect` because signcode prompts for
-    # secret key password without option for supplying it otherwise.
     aws s3 cp s3://mapbox/mapbox-studio/certs/authenticode.pvk authenticode.pvk
     aws s3 cp s3://mapbox/mapbox-studio/certs/authenticode.spc authenticode.spc
 
-    function winsign() {
-        echo "
-            spawn signcode \
-            -spc authenticode.spc \
-            -v authenticode.pvk \
-            -a sha1 -$ commercial \
-            -n Mapbox\ Studio \
-            -i https://www.mapbox.com/ \
-            -t http://timestamp.verisign.com/scripts/timstamp.dll \
-            -tr 10 \
-            $1
-
-            expect \"Enter password for authenticode.pvk:\"
-            send -- \"$2\r\"
-            expect eof
-            lassign [wait] pid spawnid os_error_flag value
-            exit \$value" | expect > /dev/null && echo "Signed $1"
-    }
-
-    # sign atom.exe to avoid false positives with antivirus software
     echo "running windows signing on atom.exe"
-    winsign $build_dir/atom.exe $WINCERT_PASSWORD
+    N='Mapbox Studio' I='https://www.mapbox.com/' P=$WINCERT_PASSWORD \
+    SPC=authenticode.spc PVK=authenticode.pvk \
+    windowsign $build_dir/atom.exe
+
     rm $build_dir/atom.exe.bak
 
     # curl -Lsfo $build_dir/resources/app/vendor/vcredist_x86.exe http://download.microsoft.com/download/2/E/6/2E61CFA4-993B-4DD4-91DA-3737CD5CD6E3/vcredist_x86.exe
@@ -145,9 +115,10 @@ if [ $platform == "win32" ]; then
     rm -rf $build_dir
     mv /tmp/mapbox-studio.exe $build_dir.exe
 
-    # sign installer
     echo "running windows signing on installer"
-    winsign $build_dir.exe $WINCERT_PASSWORD
+    N='Mapbox Studio' I='https://www.mapbox.com/' P=$WINCERT_PASSWORD \
+    SPC=authenticode.spc PVK=authenticode.pvk \
+    windowsign $build_dir.exe
 
     # remove cert
     rm -f authenticode.pvk
